Split uninitialized logic from erased definition logic
These two were crammed in PruneErasedDefs even though uninitialized does not share any logic or conceptual similarities.
